Fixed problem calling dismiss and then show
If you do a `[SVProgressHUD dismiss];` and then a `[SVProgressHUD show];` the `show` is ignored because the flag to detect if the show should do something is the `alpha` color, but it is not modified yet as it is part of a `UIAnimation`. By reducing the alpha right away we can fix the issue without visible impact.